Move `binRenders2D` system
